% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config_file.R
\name{create_config_file}
\alias{create_config_file}
\title{Create a config.yaml File from Template}
\usage{
create_config_file(config_path = "config.yaml")
}
\arguments{
\item{config_path}{The path where the config file will be saved. Defaults to `"config.yaml"`.
If the file already exists at this location, the function will prompt the user not to overwrite it.}
}
\value{
A message indicating whether the config file was successfully created or if it already exists.
}
\description{
This function copies a template `config.yaml` file to the user's working directory
if it doesn't already exist. The user can then modify the file with their own configuration details.
}
\details{
This function is useful for setting up the required configuration for the `dco` package.
The template config file contains placeholders for paths, dates, and methods that need to be filled out
by the user.
}
