% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_dco_rules.R
\name{execute_dco_rules}
\alias{execute_dco_rules}
\title{Apply DCO methods from YAML File to Data Frames}
\usage{
execute_dco_rules(
  config_path,
  data_path,
  out_path,
  out_file_suffix = "_filtered.rds",
  dco
)
}
\arguments{
\item{config_path}{Character. Path to the YAML configuration file defining rules and variables.}

\item{data_path}{The path to the directory where the data files are stored.}

\item{out_path}{The path to the directory where the output files should be saved.}

\item{out_file_suffix}{The suffix to add to the output file names.}

\item{dco}{The data cutoff date.}
}
\description{
This function reads a YAML configuration file containing method instructions
and applies the specified methods to data frames stored in a given directory.
}
