% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_date_na.R
\name{make_date_na}
\alias{make_date_na}
\title{Update end date in Data Frame}
\usage{
make_date_na(.data, start_date, end_date, cutoff_date)
}
\arguments{
\item{.data}{A data frame to update.}

\item{start_date}{The name of the start date column.}

\item{end_date}{The name of the end date column.}

\item{cutoff_date}{The cut-off date in format "YYYY-MM-DD". Rows with dates before this date will be kept.}
}
\value{
An updated data frame where `end_date` values are set to `NA` if the `end_date` is greater than the cut-off date
and the cut-off date is greater than the `start_date`.
}
\description{
This function keeps the end date in the specified column if it is before the cut-off date.
}
