% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_dco.R
\name{perform_dco}
\alias{perform_dco}
\title{Apply Rules from YAML File to Data Frames}
\usage{
perform_dco(config_path)
}
\arguments{
\item{config_path}{The path to the config.yaml file.}
}
\value{
Maintains the data frame derived from the applied functions(in RDS format).
}
\description{
Reads a YAML configuration file and applies
specified DCO rules to datasets as defined in the configuration.
}
