% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DC.R
\name{TablaAdeva}
\alias{TablaAdeva}
\title{Cross-Over Desing}
\usage{
TablaAdeva(y, trat, sujetos, periodos, significancia = 0.05, data)
}
\arguments{
\item{y}{(string) nombre de la variable respuesta.}

\item{trat}{(string) nombre de la variable que representan el tratamientos.}

\item{sujetos}{(string) nombre de la variable que representan los sujeto.}

\item{periodos}{(string) nombre de la variable que representan los periodos.}

\item{significancia}{(string) nombre de la variable que representa en grado de significancia.}

\item{data}{(\code{data.frame}) Tabla de datos en formato largo con los datos
de los tratamientos, sujetos,periodos y de la variable respuesta.}
}
\value{
Devuelve una tabla en formato \code{data.frame} con los calculos correspondientes
al analisis.
}
\description{
Obtiene la Tabla de Analisis (ADEVA) para un Diseño Cross-over/Comuntativo o Cruzado.
}
\examples{
/dontrun{
#Directorio de trabajo
ruta <-"C:/Users/Usuario/Downloads/Prueba de mi paquete/DATOS 1.csv"

#------------------------------------------------------------------
# Ejemplo 1:
df<-read.cvs(ruta)
# Cargo la libreria
library(DCruzado)
TablaAdeva(y ="respuesta",trat ="trat",sujetos = "sujetos",periodos = "periodos",significancia = 0.02,data = df)

#----------------------------------------------------------------
#Ejemplo 2:
df<- data.frame(
vaca= c(1,2,3,4,1,2,3,4,1,2,3,4,1,2,3,4),
alimento= c("B","A","C","B","A","C","D","A","C","D","A","C","D","B","B","C"),
litros_leche= c(28,31,32.5,25,29,32,34,31,27,36,34,33,34,25,28,36),
periodos= c(1,1,1,1,2,2,2,2,3,3,3,3,4,4,4,4)

TablaAdeva(y ="litros_leche",trat ="alimento",sujetos = "vaca",periodos = "periodos",significancia = 0.05,data = df)
}
}
