% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_var.R
\name{update_var}
\alias{update_var}
\title{Update Variable in Data Frame}
\usage{
update_var(.data, start_date, end_date, update_var, cutoff_date)
}
\arguments{
\item{.data}{A data frame to update.}

\item{start_date}{The name of the start date column (unquoted).}

\item{end_date}{The name of the end date column (unquoted).}

\item{update_var}{The name of the column whose values should be updated (unquoted).}

\item{cutoff_date}{The cut-off date (Date or string in "YYYY-MM-DD" format).}
}
\value{
An updated data frame where values in `update_var` are changed to "Not Recovered/Not Resolved"
if the `end_date` is after the cut-off date and `start_date` is before the cut-off.
}
\description{
Updates the values of an outcome column in a data frame if the end date is after the cut-off date,
but the cut-off date is still after the start date.
}
